Refactor for php7.
