+++ composer.json
@@ -11,25 +11,27 @@
     }
   ],
   "require": {
-    "php": ">=5.3",
-    "symfony/framework-bundle": "~2.3|~3.0",
-    "psr/log": "^1.0"
+    "php": "^7.1",
+    "symfony/framework-bundle": "^3.4 || ^4.4 || ^5.0",
+    "symfony/config": "^3.4 || ^4.4 || ^5.0",
+    "symfony/console": "^3.4 || ^4.4 || ^5.0",
+    "symfony/dependency-injection": "^3.4 || ^4.4 || ^5.0",
+    "symfony/http-kernel": "^3.4 || ^4.4 || ^5.0",
+    "ext-curl": "*",
+    "ext-json": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "4.*",
-    "squizlabs/php_codesniffer": "2.*",
-    "sebastian/phpcpd": "2.*",
-    "phing/phing": "2.*",
-    "phpmd/phpmd": "2.*",
-    "phploc/phploc": "2.*",
-    "phpunit/php-code-coverage": "^2.2"
+    "phpunit/phpunit": "^7 || ^8",
+    "symfony/yaml": "^3.4 || ^4.4 || ^5.0"
   },
   "autoload": {
     "psr-4": {
-      "Fmaj\\LaposteDatanovaBundle\\": ""
+      "Fmaj\\LaposteDatanovaBundle\\": "src"
     }
   },
-  "config": {
-    "bin-dir": "bin"
+  "autoload-dev": {
+    "psr-4": {
+      "Fmaj\\LaposteDatanovaBundle\\Tests\\": "test"
+    }
   }
 }
