+++ src/Parser/ParserInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Fmaj\LaposteDatanovaBundle\Parser;
+
+interface ParserInterface
+{
+    /**
+     * @return false|array
+     */
+    public function parse(string $dataset);
+}
